refatoracao: trasnformando AuthService em classe para melhor legibilidade
